py3/requirements.txt
cryptography==41.0.4
pyjwt==2.8.0
Flask==2.3.2
jwcrypto==1.4.2
pytest==7.4.0
coverage==7.4.0
flake8==6.1.0
